优化每股分红数据

+++ licai/src/main/java/com/sjzx/entity/CombineCashFlow.java
@@ -105,9 +105,9 @@ public class CombineCashFlow extends Model<CombineCashFlow> {
     private Long cashAndCashEquivalentsEnd;
 
     /**
-     * 分红
+     * 每股分红
      */
-    private Long bonusCash;
+    private BigDecimal bonusCash;
 
     /**
      * 员工总数
